feature/#6072 add exportToCSV function to functions.js, update effected downloadFilteredDataIntoCSV function in all system admin rool screens
